+++ d60.Cirqus/Config/Configurers/CommandProcessorConfigurationBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Configuration;
 using System.Linq;
 using d60.Cirqus.Aggregates;
 using d60.Cirqus.Commands;
@@ -47,30 +48,43 @@ namespace d60.Cirqus.Config.Configurers
             FillInDefaults();
 
             var resolutionContext = _container.CreateContext();
-
-            var eventStore = resolutionContext.Get<IEventStore>();
-            var aggregateRootRepository = resolutionContext.Get<IAggregateRootRepository>();
-            var eventDispatcher = resolutionContext.Get<IEventDispatcher>();
-            var serializer = resolutionContext.Get<IDomainEventSerializer>();
-            var commandMapper = resolutionContext.Get<ICommandMapper>();
-            var domainTypeMapper = resolutionContext.Get<IDomainTypeNameMapper>();
-
-            var commandProcessor = new CommandProcessor(eventStore, aggregateRootRepository, eventDispatcher, serializer, commandMapper, domainTypeMapper);
-
-            // end the resolution context and dispose burdens when command processor is disposed
-            commandProcessor.Disposed += resolutionContext.Dispose;
-
-            resolutionContext.GetAll<Action<Options>>()
-                .ToList()
-                .ForEach(action => action(commandProcessor.Options));
-
-            commandProcessor.Initialize();
+            var commandProcessor = resolutionContext.Get<ICommandProcessor>();
 
             return commandProcessor;
         }
 
         void FillInDefaults()
         {
+            if (_container.HasService<ICommandProcessor>(checkForPrimary: true))
+            {
+                throw new ConfigurationErrorsException("Cannot register the real CommandProcessor because the configuration container already contains a primary registration for ICommandProcessor");
+            }
+
+            _container.Register<ICommandProcessor>(context =>
+            {
+                var eventStore = context.Get<IEventStore>();
+                var aggregateRootRepository = context.Get<IAggregateRootRepository>();
+                var eventDispatcher = context.Get<IEventDispatcher>();
+                var serializer = context.Get<IDomainEventSerializer>();
+                var commandMapper = context.Get<ICommandMapper>();
+                var domainTypeMapper = context.Get<IDomainTypeNameMapper>();
+
+                var options = new Options();
+
+                context.GetAll<Action<Options>>()
+                    .ToList()
+                    .ForEach(action => action(options));
+
+                var commandProcessor = new CommandProcessor(eventStore, aggregateRootRepository, eventDispatcher, serializer, commandMapper, domainTypeMapper, options);
+
+                // end the resolution context and dispose burdens when command processor is disposed
+                commandProcessor.Disposed += context.Dispose;
+
+                commandProcessor.Initialize();
+
+                return commandProcessor;
+            });
+
             if (!_container.HasService<IAggregateRootRepository>(checkForPrimary: true))
             {
                 _container.Register<IAggregateRootRepository>(context =>